move matrix logic into matrix view to support flat view
