+++ app/src/app/views/stix/matrix/matrix-side/matrix-side.component.ts
@@ -4,8 +4,7 @@ import { Matrix } from 'src/app/classes/stix/matrix';
 import { StixObject } from 'src/app/classes/stix/stix-object';
 import { Tactic } from 'src/app/classes/stix/tactic';
 import { Technique } from 'src/app/classes/stix/technique';
-import { FilterGroup } from 'src/app/components/stix/stix-list/stix-list.component';
-import { Paginated, RestApiConnectorService } from 'src/app/services/connectors/rest-api/rest-api-connector.service';
+import { Paginated } from 'src/app/services/connectors/rest-api/rest-api-connector.service';
 
 @Component({
   selector: 'app-matrix-side',
@@ -17,52 +16,21 @@ export class MatrixSideComponent implements OnInit {
 
   private _idToLabel: Map<string, string>;
   public data$: Observable<Paginated<StixObject>>;
-  public filterOptions: FilterGroup[] = [];
-  public searchQuery: string = "";
-  public filter: string[] = [];
-  public totalObjectCount: number = 0;
   public showID = false;
   public tacticList = [];
 
-  public matrixDetails: any;
   public matrixMap: Map<string, Technique[]>;
-  constructor(
-    private restAPIConnectorService: RestApiConnectorService
-    ) {
+  constructor() {
     }
 
   ngOnInit() {
-    this.matrixMap = new Map<string, Technique[]>();
-    // order list of tactics to match ATT&Ck website
-    this.config.object.tactic_refs.forEach(item => {
-      this.tacticList.push(this.config.tacticList.find(e => e.stixID === item))
-    });
-    this.tacticList.forEach((tactic) => {
-      let itemsCopy;
-      let subscription = this.restAPIConnectorService.getTechniquesInTactic(tactic.stixID, tactic.modified).subscribe({
-        next: (items) => {
-          // alphabetically order subtechniques
-          itemsCopy = items.sort((a,b) => a.name.toLowerCase().localeCompare(b.name.toLowerCase()))
-          this.matrixMap.set(tactic.stixID, itemsCopy)
-        },
-        complete: () => {
-          subscription.unsubscribe();
-        }
-      })
-
-    })
+    this.matrixMap = this.config.matrixMap;
+    this.tacticList = this.config.tacticList;
   }
   public onToggleSubtechniquesVisible(technique: Technique) {
     technique.show_subtechniques = !technique.show_subtechniques;
   }
-  public toggleAllSubtechniquesVisible(value: boolean) {
-    for (let entry of this.matrixMap.entries()) {
-      for (let technique of entry[1])
-      {
-        technique.show_subtechniques = value
-      }
-    }
-  }
+
 
 /**
  * Get a human readable label for the given object
@@ -92,7 +60,7 @@ public getLabel(stixID: string): string {
 
 
 export interface MatrixBaseConfig {
+  matrixMap: Map<string, Technique[]>;
   field: string;
   tacticList: Tactic[];
-  object: Matrix;
 }
