+++ app/src/app/views/stix/matrix/matrix-view/matrix-view.component.ts
@@ -3,6 +3,8 @@ import { Matrix } from 'src/app/classes/stix/matrix';
 import { StixViewPage } from '../../stix-view-page';
 import { RestApiConnectorService } from 'src/app/services/connectors/rest-api/rest-api-connector.service';
 import { StixObject } from 'src/app/classes/stix/stix-object';
+import { Tactic } from 'src/app/classes/stix/tactic';
+import { Technique } from 'src/app/classes/stix/technique';
 import { AuthenticationService } from 'src/app/services/connectors/authentication/authentication.service';
 
 @Component({
@@ -13,18 +15,40 @@ import { AuthenticationService } from 'src/app/services/connectors/authenticatio
 export class MatrixViewComponent extends StixViewPage implements OnInit {
     public all_tactics: Array<StixObject>;
 
-    public view: String = "side";
+    public view: String = "flat";
     public get matrix(): Matrix { return this.config.object as Matrix; }
+
+    public tacticList = [];
+    public matrixMap: Map<string, Technique[]>;
 
     constructor(private restAPIConnectorService: RestApiConnectorService, authenticationService: AuthenticationService) {
         super(authenticationService);
     }
 
     ngOnInit() {
+      this.matrixMap = new Map<string, Technique[]>();
+
         if (!this.config.hasOwnProperty('showRelationships') || this.config.showRelationships) {
             let subscription = this.restAPIConnectorService.getAllTactics().subscribe({
                 next: (all_tactics) => {
-                    this.all_tactics = all_tactics.data;
+                  this.all_tactics = all_tactics.data;
+                  // sort tactics according to the order of attack website
+                  this.matrix.tactic_refs.forEach(item => {
+                    this.tacticList.push(all_tactics.data.find(e => e.stixID === item))
+                  });
+                  this.tacticList.forEach((tactic) => {
+                    let itemsCopy;
+                    let subscription = this.restAPIConnectorService.getTechniquesInTactic(tactic.stixID, tactic.modified).subscribe({
+                      next: (items) => {
+                        // alphabetically order subtechniques
+                        itemsCopy = items.sort((a,b) => a.name.toLowerCase().localeCompare(b.name.toLowerCase()))
+                        this.matrixMap.set(tactic.stixID, itemsCopy)
+                      },
+                      complete: () => {
+                        subscription.unsubscribe();
+                      }
+                    })
+                  })
                 },
                 complete: () => { subscription.unsubscribe(); } //prevent memory leaks
             })
@@ -34,7 +58,12 @@ export class MatrixViewComponent extends StixViewPage implements OnInit {
         }
 
     }
-    toggleAllSubtechniquesVisible(i: boolean) {
-
+    public toggleAllSubtechniquesVisible(value: boolean) {
+      for (let entry of this.matrixMap.entries()) {
+        for (let technique of entry[1])
+        {
+          technique.show_subtechniques = value
+        }
+      }
     }
 }
